pyproject.toml
[project]
name = "brainaudio"
version = "0.1.0"
description = "Code for building algorithms for speech neuroprostheses."
readme = "README.md"
authors = [
    { name = "ebrahimfeghhi", email = "efeghhi@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "edit-distance>=1.0.7",
    "einops>=0.8.1",
    "g2p-en>=2.1.0",
    "h5py>=3.14.0",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pyyaml>=6.0.3",
    "scipy>=1.16.2",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "wandb>=0.22.1",
]

[build-system]
requires = ["uv_build>=0.8.23,<0.9.0"]
build-backend = "uv_build"
